+ Add Scope, FunctionProvider, Extension
